requirements.txt
Django>=6.0,<6.1
Pillow>=11.0
djangorestframework>=3.17.0
django-filter>=25.0
drf-spectacular>=0.29.0
python-dotenv>=1.0.0
